A308 V1.37 20230426 (曾) Server在A308模式中，即便接著市電也會進入休眠 2. 發送RS485 TX訊息前不論Uart是否在忙錄中都將狀態重置以發送訊息(原流程為直接跳過不發送)

+++ app.c
@@ -76,7 +76,7 @@ const char* SensorClassStr[]=
     "NO Sensor","Auto Scan","PZEM","Visual Sensor","DC600","FTM94","BTM-G6","BTM485"
 };
 extern void DebugShowSetting();
-extern uint16 IndexToBaudrate(uint8);
+extern uint32 IndexToBaudrate(uint8);
 //***************************************************************************
 // Initial BLE Stack, IO port,
 //
